+++ ci/build-dmg.sh
@@ -104,8 +104,8 @@ cp -av /usr/local/Cellar/libmagic/*/share/misc galacteek.app/Contents/Resources/
 cp -av ${GITHUB_WORKSPACE}/packaging/macos/zbar/*.dylib galacteek.app/Contents/Resources/lib
 cp -av ${GITHUB_WORKSPACE}/packaging/macos/jpeg/*.dylib galacteek.app/Contents/Resources/lib
 
-cp /usr/local/Cellar/openssl@1.1/1.1.1l_1/lib/libssl*.dylib galacteek.app/Contents/Resources/lib
-cp /usr/local/Cellar/openssl@1.1/1.1.1l_1/lib/libcrypto*.dylib galacteek.app/Contents/Resources/lib
+cp /usr/local/Cellar/openssl@1.1/1.1.1m/lib/libssl*.dylib galacteek.app/Contents/Resources/lib
+cp /usr/local/Cellar/openssl@1.1/1.1.1m/lib/libcrypto*.dylib galacteek.app/Contents/Resources/lib
 
 # create entry script for galacteek
 cat > galacteek.app/Contents/MacOS/galacteek <<\EAT